A temporary workaround for the PROPFIND issue.
There is an issue in ntlmrelayx's HTTP server preventing the WebDAV
server from working as intended. This issue is caused by changes to the
HTTP relay server and target processor. This change ports the target
processor from 95f95915e54cc743c78d8683b14f9a3e39b3c0c7 forward.

This is intended as a quick and dirty workaround and likely breaks other
ntlmrelayx functionality modified for compatibility with the newer
target manager.

There appears to be a problem with how the HTTP relay server handles
targets that may result in a target being consumed for every HTTP
request, regardless of the request's relevance to an attack. This will
be investigated and addressed at a later time.